refactor(e2e): Optimize Import Paths and LoginPage Functionality
