+++ e2e/dotcms-e2e-node/frontend/src/tests/login/login.spec.ts
@@ -1,5 +1,6 @@
 import { test, expect } from "@playwright/test";
 import { admin1, wrong1, wrong2 } from "./credentialsData";
+import { LoginPage } from "@pages";
 
 const validCredentials = [
   { username: admin1.username, password: admin1.password }, // admin user
@@ -10,16 +11,13 @@ const validCredentials = [
  */
 validCredentials.forEach(({ username, password }) => {
   test(`Login with Valid Credentials: ${username}`, async ({ page }) => {
-    await page.goto("/dotAdmin");
+    const loginPage = new LoginPage(page);
+    await loginPage.login(username, password);
 
-    await page.fill('input[id="inputtext"]', username);
-    await page.fill('input[id="password"]', password);
-    await page.getByTestId("submitButton").click();
-
-    // Assertion and further test steps
-    await expect(
-      page.getByRole("link", { name: "Getting Started" }),
-    ).toBeVisible();
+    const gettingStartedLocator = page.getByRole("link", {
+      name: "Getting Started",
+    });
+    await expect(gettingStartedLocator).toBeVisible();
   });
 });
 
@@ -37,13 +35,10 @@ invalidCredentials.forEach((credentials) => {
   }) => {
     const { username, password } = credentials;
 
-    await page.goto("/dotAdmin");
+    const loginPage = new LoginPage(page);
+    await loginPage.login(username, password);
 
-    await page.fill('input[id="inputtext"]', username);
-    await page.fill('input[id="password"]', password);
-    await page.getByTestId("submitButton").click();
-
-    // Assertion and further test steps
-    await expect(page.getByTestId("message")).toBeVisible({ timeout: 30000 });
+    const errorMessageLocator = page.getByTestId("message");
+    await expect(errorMessageLocator).toBeVisible({ timeout: 30000 });
   });
 });
