+++ e2e/dotcms-e2e-node/frontend/src/tests/newEditContent/fields/siteOrFolderField.spec.ts
@@ -1,10 +1,6 @@
 import { test, expect } from "@playwright/test";
 import { faker } from "@faker-js/faker";
-import { ListingContentTypesPage } from "@pages/listingContentTypes.page";
-import { ContentTypeFormPage } from "@pages/contentTypeForm.page";
-import { NewEditContentFormPage } from "@pages/newEditContentForm.page";
-import { ListingContentPage } from "@pages/listingContent.page";
-import { LoginPage } from "@pages/index";
+import { ListingContentTypesPage, ContentTypeFormPage, NewEditContentFormPage, ListingContentPage, LoginPage } from "@pages";
 import { createDefaultContentType } from "@data/defaultContentType";
 
 const contentTypeName = faker.lorem.word().toLocaleLowerCase();
